set up generic mp4box parse tests
