+++ .gitattributes
@@ -1,2 +1,3 @@
 * text=auto eol=lf
+*.mp4 filter=lfs diff=lfs merge=lfs -text
 Cargo.lock linguist-generated=false
