+++ tests/README.md
@@ -1,0 +1,9 @@
+# Tests
+
+Execution of tests requires [git-lfs](https://git-lfs.github.com/) to download sample data.
+Ensure that all video samples are pushed as lfs files!
+
+## mp4box comparision tests
+
+These tests compare the output of `mp4box` against the output of this library.
+In order to run these tests, you'll need to install `node` (and have `node` in your `PATH`).
