+++ tests/samples/.gitignore
@@ -1,0 +1,3 @@
+*.segments
+*.json
+*.bin
